requirements.txt
streamlit>=1.29.0
pandas>=2.0.0
pdfplumber>=0.10.2
easyocr>=1.7.1
opencv-python-headless>=4.8.0.76
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
fpdf>=1.7.2
Pillow>=10.0.0
openpyxl>=3.1.2
torch>=2.0.0
transformers>=4.36.2